move app to src directory
